requirements.txt
alembic==1.9.4
anyio==3.6.2
click==8.1.3
colorama==0.4.6
fastapi==0.92.0
fastapi-utils==0.2.1
greenlet==2.0.2
h11==0.14.0
httptools==0.5.0
idna==3.4
install==1.3.5
Mako==1.2.4
MarkupSafe==2.1.2
pydantic==1.10.5
python-dotenv==0.21.1
PyYAML==6.0
sniffio==1.3.0
SQLAlchemy==1.4.46
starlette==0.25.0
typing-extensions==4.5.0
uvicorn==0.20.0
watchfiles==0.18.1
websockets==10.4
python-dateutil~=2.8.2
pytest==7.2.1
freezegun==1.2.2